Add widget LvmVgDevicesSelector
